requirements.txt
customtkinter>=5.2.0
watchdog>=3.0.0
pystray>=0.19.0
Pillow>=10.0.0
httpx>=0.25.0
